% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rep.R
\name{email_rep}
\alias{email_rep}
\title{Get email reputation}
\usage{
email_rep(email, ...)
}
\arguments{
\item{email}{the email to get rep for}

\item{...}{passed on to \code{\link[httr:GET]{httr::GET()}}}
}
\description{
Get email reputation
}
\examples{
email_rep("bill@microsoft.com")
}
\references{
\url{https://emailrep.io/}; \url{https://github.com/sublime-security/emailrep.io};
\url{https://emailrep.io/docs/#emailrep-alpha-api}
}
