% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailrep-package.R
\docType{package}
\name{emailrep}
\alias{emailrep}
\alias{emailrep-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/emailrep}
\item BugReports: \url{https://gitlab.com/hrbrmstr/emailrep/issues}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/emailrep}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/emailrep/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
