% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailrep-package.R
\docType{package}
\name{emailrep}
\alias{emailrep}
\alias{emailrep-package}
\title{Check Email Address Reputation Against the 'EmailRep' API}
\description{
EmailRep is a system of crawlers, scanners and enrichment services
that collects data on email addresses, domains, and internet personas.
EmailRep uses hundreds of data points from social media profiles,
professional networking sites, dark web credential leaks, data breaches,
phishing kits, phishing emails, spam lists, open mail relays, domain age and
reputation, deliverability, and more to predict the risk of an email address.
Tools are provided to gather reputation information for emails.
}
\references{
\url{https://emailrep.io/}; \url{https://github.com/sublime-security/emailrep.io};
\url{https://emailrep.io/docs/#emailrep-alpha-api}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/emailrep}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/emailrep/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
